Begin moving over the old unit tests to Jest / TypeScript
