+++ packages/cli/test/tsconfig.json
@@ -1,4 +1,4 @@
 {
   "extends": "../tsconfig.json",
-  "include": ["*.test.ts"]
+  "include": ["**/*.test.ts"]
 }
